+++ generators/lcg32prime.c
@@ -49,7 +49,7 @@ static inline uint64_t get_bits_raw(void *state)
 static void *create(const CallerAPI *intf)
 {
     Lcg32State *obj = intf->malloc(sizeof(Lcg32State));
-    obj->x = (intf->get_seed64() >> 32) | 0x1;
+    obj->x = (uint32_t) (intf->get_seed64() >> 32) | 0x1;
     return (void *) obj;
 }
